requirement.txt
Django
djangorestframework
django-cors-headers
drf-yasg
django-environ
mysqlclient
celery